[IMP]Improved code
