+++ business_requirement/models/business.py
@@ -4,6 +4,7 @@
 
 from openerp import api, fields, models, _
 from openerp.exceptions import except_orm
+from openerp.exceptions import Warning as UserError
 
 
 class BusinessRequirement(models.Model):
@@ -314,6 +315,74 @@ class BusinessRequirement(models.Model):
             content_subtype=content_subtype, **kwargs
         )
         return res
+
+    @api.model
+    def read_group(self, domain, fields, groupby, offset=0,
+                   limit=None, orderby=False, lazy=True):
+        if groupby and groupby[0] == "state":
+            states = self.env['business.requirement'].\
+                fields_get(['state']).get('state').get('selection')
+            read_group_all_states = [{'__context': {'group_by': groupby[1:]},
+                                      '__domain': domain + [('state', '=',
+                                                             state_value)],
+                                      'state': state_value,
+                                      'state_count': 0}
+                                     for state_value, state_name in states]
+            # Get standard results
+            read_group_res = super(BusinessRequirement, self).\
+                read_group(domain, fields, groupby, offset=offset,
+                           limit=limit, orderby=orderby)
+            # Update standard results with default results
+            result = []
+            for state_value, state_name in states:
+                res = filter(lambda x: x['state'] == state_value,
+                             read_group_res)
+                if not res:
+                    res = filter(lambda x: x['state'] == state_value,
+                                 read_group_all_states)
+                res[0]['state'] = [state_value, state_name]
+                result.append(res[0])
+            return result
+        else:
+            return super(BusinessRequirement, self).\
+                read_group(domain, fields, groupby,
+                           offset=offset, limit=limit, orderby=orderby)
+
+    @api.multi
+    def write(self, vals):
+        for r in self:
+            if vals.get('state'):
+                ir_obj = self.env['ir.model.data']
+                br_xml_id = ir_obj.\
+                    get_object('business_requirement',
+                               'group_business_requirement_manager')
+                user = self.env['res.users']
+                grps = [grp.id for grp in user.browse(self._uid).groups_id]
+                date = fields.Datetime.now()
+                if vals['state'] == 'confirmed':
+                    vals.update({'confirmed_id': user,
+                                 'confirmation_date': date})
+                if vals['state'] == 'draft':
+                    vals.update({'confirmed_id': False,
+                                 'approved_id': False,
+                                 'confirmation_date': False,
+                                 'approval_date': False
+                                 })
+                if vals['state'] == 'approved':
+                    if br_xml_id.id in grps:
+                        vals.update({'approved_id': user,
+                                     'approval_date': date})
+                    else:
+                        raise UserError(_('You can only move to the '
+                                          'following stage: draft/confirmed'
+                                          '/cancel/drop.'))
+                if vals['state'] in ('stakeholder_approval', 'in_progress',
+                                     'done'):
+                    if br_xml_id.id not in grps:
+                        raise UserError(_('You can only move to the'
+                                          'following stage: draft/'
+                                          'confirmed/cancel/drop.'))
+            return super(BusinessRequirement, self).write(vals)
 
 
 class BusinessRequirementCategory(models.Model):
